Fix a sandwich bug and add css
There was a problem with sandwich.html's title, it was appearing inline
on the body as if it were loose text, making it appear as sandwich
header with a "sandwich" above it. The code seemed the same as cereal
and lasagna's, ended up copy pasting lasagna's on top and just changing
the heading to sandwich's and that fixed it.

WEIRD!